Rename tests_requre test suite and zuul job
The suite contains also other tests (not only ones using requre)
so the name should make clear that is's about tests using the deployed openshift and running database.
