+++ tests_openshift/openshift_integration/test_copr.py
@@ -4,7 +4,7 @@
 import json
 import unittest
 
-from tests_requre.openshift_integration.base import PackitServiceTestCase, DATA_DIR
+from tests_openshift.openshift_integration.base import PackitServiceTestCase, DATA_DIR
 
 
 def pr_event():
