+++ tests_openshift/openshift_integration/test_data/test_fedpkg/tests_openshift.openshift_integration.test_fedpkg.FedPkg.test_fedpkg_clone.yaml.static_tmp_1_1.tar.xz
